CIN-25: CastAndCrew 리액트 포탈 도입
